+++ src/entity/Cliente.java
@@ -4,7 +4,7 @@ import java.util.logging.Logger;
 
 public class Cliente extends Usuario{
 
-    public Cliente(Long codigo, String nome, String email, String senha, String tipo) {
+    public Cliente(int codigo, String nome, String email, String senha, String tipo) {
         super(codigo, nome, email, senha, tipo);
     }
 }
