+++ src/entity/Profissional.java
@@ -1,7 +1,7 @@
 package entity;
 
 public class Profissional extends Usuario{
-    public Profissional(Long codigo, String nome, String email, String senha, String tipo) {
+    public Profissional(int codigo, String nome, String email, String senha, String tipo) {
         super(codigo, nome, email, senha, tipo);
     }
 }
